requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic-settings>=2.5.0
celery[redis]>=5.4.0
redis>=5.0.0
httpx>=0.27.0
python-multipart>=0.0.12
python-magic>=0.4.27
structlog>=24.4.0
prometheus-fastapi-instrumentator>=7.0.0
python-dotenv>=1.0.1

# Testing
pytest>=9.0.0
pytest-asyncio>=0.24.0
pytest-mock>=3.14.0
